add casts to suppress compiler warnings
git-svn-id: https://svn.r-project.org/R/trunk@73458 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/summary.c
@@ -487,7 +487,7 @@ static R_INLINE SEXP complex_mean(SEXP x)
 	}
 	s += t/n; si += ti/n;
     }
-    Rcomplex val = { s, si };
+    Rcomplex val = { (double)s, (double)si };
     return ScalarComplex(val);
 }
